+++ src/drbd/utilities/Tools.java
@@ -2077,9 +2077,7 @@ public final class Tools {
         }
     }
 
-    /**
-     * Converts value in kilobytes.
-     */
+    /** Converts value in kilobytes. */
     public static String convertKilobytes(final String kb) {
         if (isNumber(kb)) {
             final double k = Integer.parseInt(kb);
@@ -2103,10 +2101,8 @@ public final class Tools {
         return kb;
     }
 
-    /**
-     * Converts value with unit to kilobites.
-     */
-    public static int convertToKilobytes(final String value) {
+    /** Converts value with unit to kilobites. */
+    public static long convertToKilobytes(final String value) {
         final Object[] v = Tools.extractUnit(value);
         if (v.length == 2 && Tools.isNumber((String) v[0])) {
             int num = Integer.parseInt((String) v[0]);
